Add array-converting functionality to Ui class
Ui class does not have any function.

Methods that assist functionality (such as converting an array to string) should be handled by the Ui class.

Let's bring the tasklistToString() method from the Command class to the Ui class. This will follow OOP principles better.
