+++ data/tasks.txt
@@ -4,4 +4,3 @@ E | 0 | baby shower | 24/08/2021 5:30PM
 T | 1 | join sports club
 T | 1 | test something
 D | 0 | wish bday | 04/10/2021 11:59PM
-T | 0 | something
