+++ src/main/java/duke/Duke.java
@@ -1,27 +1,14 @@
 package duke;
 
-import javafx.application.Application;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import javafx.geometry.Pos;
-import javafx.scene.Node;
-import javafx.scene.control.Label;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.HBox;
-import javafx.scene.Scene;
 import javafx.scene.control.Button;
-import javafx.scene.control.Label;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.Region;
 import javafx.scene.layout.VBox;
-import javafx.stage.Stage;
+import javafx.scene.Scene;
+
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
-import java.util.Scanner;
 
 import duke.commands.Command;
 import duke.parser.Parser;
@@ -34,15 +21,6 @@ import duke.ui.Ui;
  * Initializes the application and starts the interaction with the user.
  */
 public class Duke {
-    // JavaFX GUI Elements
-    private Image user = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
-    private Image duke = new Image(this.getClass().getResourceAsStream("/images/DaDuke.png"));
-    private ScrollPane scrollPane;
-    private VBox dialogContainer;
-    private TextField userInput;
-    private Button sendButton;
-    private Scene scene;
-
     // Fields for chatbot functionality
     public static final String FILEPATH = "data/tasks.txt";
     private TaskList tasks;
@@ -59,7 +37,6 @@ public class Duke {
         try {
             tasks = new TaskList(storage.load());
         } catch (DukeException e) {
-            ui.showLoadingError();
             tasks = new TaskList();
         }
 
@@ -71,7 +48,6 @@ public class Duke {
         try {
             tasks = new TaskList(storage.load());
         } catch (DukeException e) {
-            ui.showLoadingError();
             tasks = new TaskList();
         } catch (IOException e) {
             System.out.println("Critical failure while reading io");
@@ -80,8 +56,6 @@ public class Duke {
 
     public String getResponse(String input) {
         try {
-            //String fullCommand = ui.readCommand();
-            ui.showLine(); // show the divider line ("_______")
             Command c = Parser.parse(input);
             c.execute(tasks, ui, storage);
             return c.getCommandOutput();
