+++ src/main/java/duke/ui/Ui.java
@@ -1,5 +1,9 @@
 package duke.ui;
 
+import duke.tasks.Task;
+
+import java.util.ArrayList;
+
 /**
  * Text UI of the application.
  */
@@ -8,15 +12,19 @@ public class Ui {
 
     }
 
-    public static void showLoadingError() {
-
-    }
-
-    public void readCommand() {
-
-    }
-
-    public void showLine() {
-
+    /**
+     * Constructs a message containing all tasks in the ArrayList.
+     *
+     * @param list list of tasks
+     * @return ArrayList in String format
+     */
+    public static String arrayToString(ArrayList<Task> list) {
+        String answer = "";
+        int counter = 1;
+        for (Task item : list) {
+            answer += String.format("%d: %s\n", counter, item.toString());
+            counter++;
+        }
+        return answer;
     }
 }
